feat, refactor: adds adtl. functionality to record checking, adds helper function.py, for valid date checking, refactors user creation methodology
Due to limitations with using the original array, A new list is created as we loop through our excel list. This is to prevent out of bounds errors at time of user creation. if at any point an error arises, the user array will no longer be used and users will not be added