Feat: TeamDisplay basic functionality added
Created a view, url, and a HTML page to recieve a team ID and print out
basic information about said team along with the members on given team.
Projects the team is on still needs to be added, along with reformating
the output page to make it look nice, finally navigation bar needs to be added.