feat: Removed notification indicator, polished directory.html
In nav.html and search.html, removed the notification indicator in the navigation bar.

In directory.html and directory.css, the directory page has now been completed and is polished. If the list of teams, users, or projects goes beyond the height of these lists, a scrollbar will appear for ease of viewing.